fix(macos): pin testClampRad TA-short checks to the defined int32 conversion
The SYNC-003 gate (testClampRad) still used raw float->short casts,
which are UB once the value leaves short's range. On arm64 -O2 clang
lowers them to fcvtzs-with-saturation, so the near-full-turn boundary
check (short(ClampRad(nextafter(TWOPI,-inf))*RAD2TAANG) == -1) failed
with 65535 != -1 - even though the engine sweep itself (8a041014a2 +
7104ab1f76) already pins the defined float->int32->short form at every
synced site, matching the fleet x86 cvttss2si-then-narrow behavior.

Rewrite the 7 TA-short checks to go through int32 exactly like the
swept engine sites (CobInstance RadAngleToCobShort, IPathController,
Ground/HoverAirMoveType, LuaSyncedMoveCtrl/LuaSyncedRead). All
in-bounds results are unchanged; only the UB boundary case is pinned
to the fleet value.

Adds the ACDD spec for the sweep (SYNC-003, done) with the @sync-gate
component tag. ctest in build-mac: 28/30 pass (testCreg and
testUnitSync fail for pre-existing environmental reasons: missing
spring-headless binary and missing base/springcontent.sdz content
respectively - unrelated to this change).

AI-assisted: Claude (rig) performed the ACDD discovery, root-cause
analysis (compiler lowering verified with standalone repros), test
edit, and spec bookkeeping.

+++ test/engine/System/testClampRad.cpp
@@ -1,4 +1,13 @@
 /* This file is part of the Spring engine (GPL v2 or later), see LICENSE.html */
+/**
+ * Feature: spec/features/synced-code-float-short-ub-sweep-matches-fleet-gcc-x86-behavior.feature
+ *
+ * Validates the synced-code float→short UB sweep (SYNC-003): the TA-short
+ * expectations must match the fleet's gcc-x86 behavior (float→int32→short,
+ * i.e. cvttss2si-then-narrow). The out-of-range raw float→short cast the
+ * gate used before this fix is UB and produced platform-dependent results
+ * on arm64 (-O2: fcvtzs saturation → 65535 instead of -1).
+ */
 
 #include <cmath>
 #include <numeric>
@@ -43,10 +52,18 @@ TEST_CASE("ClampRad")
 	CHECK_FALSE(std::signbit(ClampRad(0.0f)));
 
 	// Test TAANG2RAD conversion to short for [0, 2pi)
-	CHECK(static_cast<short>(ClampRad(0.0f) * RAD2TAANG) == short(0));
-	CHECK(static_cast<short>(ClampRad(+std::nextafterf(math::TWOPI, -std::numeric_limits<float>::infinity())) * RAD2TAANG) == short(-1));
-	CHECK(static_cast<short>(ClampRad(+std::nextafterf(       0.0f, +std::numeric_limits<float>::infinity())) * RAD2TAANG) == short( 0));
-	CHECK(static_cast<short>(ClampRad(+std::nextafterf(TAANG2RAD  , +std::numeric_limits<float>::infinity())) * RAD2TAANG) == short(+1));
+	// The float->int->short form mirrors the engine's synced-code sweep (see
+	// CobInstance.cpp RadAngleToCobShort, SYNC-003): a raw float->short cast
+	// is UB once the value leaves short's range, and arm64 -O2 lowers it to
+	// fcvtzs-with-saturation (65535.996 -> 65535) instead of the fleet's
+	// x86 cvttss2si-then-narrow wrap (-> -1).
+	// @step Given the gate test drives ClampRad outputs through the port's defined float→int32→short conversion
+	// @step When the boundary angles (0, nextafter(TWOPI, -inf), nextafter(0, +inf), nextafter(TAANG2RAD, +inf)) are converted to TA-short
+	// @step Then the converted values equal the fleet's gcc-x86 results: 0, -1 (the 65536 wrap), 0, +1
+	CHECK(static_cast<short>(static_cast<int>(ClampRad(0.0f) * RAD2TAANG)) == short(0));
+	CHECK(static_cast<short>(static_cast<int>(ClampRad(+std::nextafterf(math::TWOPI, -std::numeric_limits<float>::infinity())) * RAD2TAANG)) == short(-1));
+	CHECK(static_cast<short>(static_cast<int>(ClampRad(+std::nextafterf(       0.0f, +std::numeric_limits<float>::infinity())) * RAD2TAANG)) == short( 0));
+	CHECK(static_cast<short>(static_cast<int>(ClampRad(+std::nextafterf(TAANG2RAD  , +std::numeric_limits<float>::infinity())) * RAD2TAANG)) == short(+1));
 }
 
 TEST_CASE("ClampRadPi")
@@ -76,8 +93,12 @@ TEST_CASE("ClampRadPi")
 	CHECK_FALSE(std::signbit(ClampRadPi(0.0f)));
 
 	// Test TAANG2RAD conversion to short for [-pi, pi)
-	CHECK(static_cast<short>(ClampRadPi(-(math::PI)) * RAD2TAANG) == short(-32768));
-	CHECK(static_cast<short>(ClampRadPi(+std::nextafterf(math::PI, -std::numeric_limits<float>::infinity())) * RAD2TAANG) == short(32767));
-	CHECK(static_cast<short>(ClampRadPi((math::PI)) * RAD2TAANG) == short(-32768));
+	// Same int-intermediate form as the [0, 2pi) block above (SYNC-003).
+	// @step Given the gate test drives ClampRadPi outputs through the port's defined float→int32→short conversion
+	// @step When the boundary angles (-PI, nextafter(PI, -inf), PI) are converted to TA-short
+	// @step Then the converted values equal the fleet's gcc-x86 results: -32768 (exactly at the short minimum), 32767 (truncation of 32767.998, not 32768), -32768
+	CHECK(static_cast<short>(static_cast<int>(ClampRadPi(-(math::PI)) * RAD2TAANG)) == short(-32768));
+	CHECK(static_cast<short>(static_cast<int>(ClampRadPi(+std::nextafterf(math::PI, -std::numeric_limits<float>::infinity())) * RAD2TAANG)) == short(32767));
+	CHECK(static_cast<short>(static_cast<int>(ClampRadPi((math::PI)) * RAD2TAANG)) == short(-32768));
 }
